Added styles for bot-head
Added styles for bot-head such as dividing the bot-head into two parts, giving each of them two seperate colors. But there is problem with the ul li. They stay in the top-part. I need them to be vertically centered within the bot-head. 

In addition, I removed the outline from mid-head.